moved previous skeletons in main component files
